refactor(ffi-py): split lib.rs into per-feature modules
Pure refactor: 1346-line lib.rs becomes a 214-line shim plus seven
per-feature modules per issue #36, mirroring the bridge crate's
sub-project decomposition. The Python-facing surface
(secretary_ffi_py.* classes / functions / exceptions) is unchanged;
all `m.add_class::<...>` and `m.add(...)` registrations stay in the
single #[pymodule] entrypoint to keep the Python API contract pinned
to one place.

- lib.rs (214)        — module docs + #[pymodule] registration +
                         smoke (add, version, version_py) + tests
- errors.rs (116)     — 17 create_exception! macros + the two
                         FfiX → PyErr translators + uuid_array_or_value_error
- identity.rs (56)    — UnlockedIdentity pyclass (shared dependency)
- unlock.rs (228)     — bytes-in B.2/B.3a/B.3b: open_with_password,
                         open_with_recovery, create_vault, MnemonicOutput,
                         CreateVaultOutput
- vault.rs (301)      — folder-in B.4a: open_vault_with_password,
                         open_vault_with_recovery, BlockSummary,
                         OpenVaultManifest, OpenVaultOutput
- record.rs (222)     — B.4b: read_block + FieldHandle + Record +
                         BlockReadOutput
- save.rs (207)       — B.4c: save_block + BlockInput + RecordInput +
                         FieldInput + FieldInputValue
- share.rs (48)       — B.4d: share_block

UnlockedIdentity::close() and MnemonicOutput's wipe accessors required
pub(crate) widening so vault::OpenVaultOutput::__exit__ can still
cascade-close them across module boundaries.

Test counts unchanged: 599 cargo passed + 9 ignored, 57 pytest passed.
